Modularizacion de ArtistInfoRepository
