+++ app/src/main/java/ayds/lisboa/songinfo/moredetails/fulllogic/ArtistInfoRepository.kt
@@ -1,11 +1,9 @@
 package ayds.lisboa.songinfo.moredetails.fulllogic
 
-import android.content.ContentValues
 import ayds.lisboa.songinfo.moredetails.fulllogic.ArtistInfo.LastFmArtistInfo
 
 interface ArtistInfoRepository{
-    fun saveArtist(artist: String, artistInfo: LastFmArtistInfo)
-    fun getArtistInfo(artist: String): LastFmArtistInfo?
+    fun getArtistInfo(artistName: String): ArtistInfo
 }
 
 internal class ArtistInfoRepositoryImpl(
@@ -14,28 +12,6 @@ internal class ArtistInfoRepositoryImpl(
 ) : ArtistInfoRepository {
 
 
-    override fun saveArtist(artist: String, artistInfo: LastFmArtistInfo) {
-        val artistValues = getArtistValues(artist, artistInfo)
-        lastFMLocalStorage.saveArtist(artistValues)
-    }
-
-    private fun getArtistValues(artist: String, artistInfo: LastFmArtistInfo): ContentValues {
-        val values = ContentValues().apply {
-            put(ARTIST, artist)
-            put(BIO_CONTENT, artistInfo.bioContent)
-            put(URL, artistInfo.url)
-            put(SOURCE, 1)
-        }
-
-        return values
-    }
-    /*
-    override fun getArtistInfo(artist: String): LastFmArtistInfo? {
-        val artistValues = arrayOf(artist)
-
-        return lastFMLocalStorage.getArtistInfo(artistValues)
-    }
-*/
     override fun getArtistInfo(artistName: String): ArtistInfo {
         var artistInfo = lastFMLocalStorage.getArtistInfo(artistName)
 
@@ -43,7 +19,7 @@ internal class ArtistInfoRepositoryImpl(
             artistInfo != null -> markArtistInfoAsSavedDB(artistInfo)
             else -> {
                 try {
-                    artistInfo = lastFMService.getArtistInfoFromService(artistName)
+                    artistInfo = lastFMService.getArtistInfo(artistName)
                     saveArtistInfoDB(artistName, artistInfo)
                 } catch (ioException: Exception) {
                     ioException.printStackTrace()
