+++ app/src/main/java/ayds/lisboa/songinfo/moredetails/fulllogic/LastFMService.kt
@@ -33,4 +33,6 @@ internal class LastFMServiceImpl: LastFMService {
     override fun getResponse(artistName: String): Response<String> {
         return lastFMAPI.getArtistInfo(artistName).execute()
     }
+
+
 }
